+++ config/config.app.php
@@ -308,9 +308,9 @@ else {
  * !Do not change anything below this line. It is protected by the license agreement!
  */
 define('APP_NAME', "TeamCal Neo");
-define('APP_VER', "5.3.3");
-define('APP_BUILD', "103");
-define('APP_DATE', "2026-07-08");
+define('APP_VER', "5.3.4");
+define('APP_BUILD', "105");
+define('APP_DATE', "2026-07-17");
 define('APP_YEAR', "2014-" . date('Y'));
 define('APP_AUTHOR', "George Lewe");
 define('APP_URL', "https://www.lewe.com");